requirements.txt
# ==============================================================
#  FinAgent - Python Dependencies
#  Install with: pip install -r requirements.txt
# ==============================================================

# --- Data Collection ---
yfinance>=0.2.40           # Yahoo Finance stock price & fundamentals
requests>=2.32.0           # HTTP client for REST APIs
httpx>=0.27.0              # Async-capable HTTP client (Alpha Vantage, NewsAPI)
beautifulsoup4>=4.12.0     # HTML parsing for web scraping
lxml>=5.2.0                # Fast XML/HTML parser backend for BeautifulSoup

# --- Data Processing ---
pandas>=2.2.0              # Core tabular data manipulation
numpy>=1.26.0              # Numerical computing
scipy>=1.13.0              # Statistical functions (Z-score, IQR helpers)

# --- Visualisation ---
matplotlib>=3.9.0          # Static chart rendering
seaborn>=0.13.0            # Statistical visualisation (heatmap, KDE)
plotly>=5.22.0             # Interactive charts
kaleido>=0.2.1             # Static image export for Plotly figures

# --- AI / LLM Integration ---
google-generativeai>=0.7.0 # Google Gemini SDK (default provider)
anthropic>=0.28.0          # Anthropic Claude SDK
openai>=1.35.0             # OpenAI GPT SDK

# --- Environment & Configuration ---
python-dotenv>=1.0.1       # Load .env variables into os.environ
python-dateutil>=2.9.0     # Accurate relative date arithmetic (relativedelta)

# --- Jupyter (optional - for notebooks/) ---
jupyter>=1.0.0
ipykernel>=6.29.0

# --- Development & Code Quality ---
black>=24.4.0              # Code formatter (PEP 8)
flake8>=7.1.0              # Linter
isort>=5.13.0              # Import sorter
pytest>=8.2.0              # Test runner
pytest-cov>=5.0.0          # Coverage reporting